Prepare enterprise repository
