--- .conan/profiles/linux_aarch64_release
@@ -1,7 +1,0 @@
-[settings]
-os=Linux
-arch=armv8
-compiler=gcc
-compiler.version=11
-compiler.libcxx=libstdc++11
-build_type=Release
