--- .conan/profiles/linux_x86_64_release
@@ -1,7 +1,0 @@
-[settings]
-os=Linux
-arch=x86_64
-compiler=gcc
-compiler.version=11
-compiler.libcxx=libstdc++11
-build_type=Release
