+++ .github/workflows/ci.yml
@@ -39,16 +39,19 @@ jobs:
   build-and-publish:
     runs-on: ubuntu-22.04
     container: ghcr.io/eclipse-velocitas/devcontainer-base-images/cpp:v0.3
-    name: "Build, Test and Lint"
+    name: "Build, test and lint"
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
-        with:
-          submodules: "recursive"
 
       - name: Fix dubious ownership
         run: |
           git config --global --add safe.directory $( pwd )
+
+      - name: Write conan credentials
+        run: |
+          echo "CONAN_REMOTE_USER=${{ secrets.BDC_ARTIFACTORY_USER }}" >> .credentials
+          echo "CONAN_REMOTE_TOKEN=${{ secrets.BDC_ARTIFACTORY_PAT }}" >> .credentials
 
       - name: Init velocitas project
         run: |
@@ -59,10 +62,10 @@ jobs:
           pip3 install -r requirements.txt
 
       - name: Install dependencies
-        run: ./install_dependencies.sh -r
+        run: velocitas exec build-system install -r
 
       - name: build
-        run: ./build.sh -r
+        run: velocitas exec build-system build -r
 
       - name: Run Unit Tests
         run: build/bin/app_utests > test-results.md
