+++ .github/workflows/localversion.yml
@@ -1,0 +1,104 @@
+# This file is maintained by velocitas CLI, do not modify manually. Change settings in .velocitas.json
+# Copyright (c) 2022-2024 Contributors to the Eclipse Foundation
+#
+# This program and the accompanying materials are made available under the
+# terms of the Apache License, Version 2.0 which is available at
+# https://www.apache.org/licenses/LICENSE-2.0.
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+name: Release local version
+
+on:
+  release:
+    types: [published, edited]
+
+  workflow_dispatch:
+  push:
+    # Run only on branches/commits and not tags
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+
+jobs:
+
+  publish-readme:
+    name: Publish README
+    runs-on: ubuntu-22.04
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Archive local setup README
+        uses: actions/upload-artifact@v4
+        with:
+          name: LOCAL README
+          path: ./.devcontainer/localversion/README.md
+
+  create-local-image:
+    name: Create local version
+    runs-on: ubuntu-22.04
+    strategy:
+      matrix:
+        platform: [linux/amd64, linux/aarch64]
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Fix dubious ownership
+        run: |
+          git config --global --add safe.directory $( pwd )
+
+      - name: Write conan credentials
+        run: |
+          echo "CONAN_REMOTE_USER=${{ secrets.BDC_ARTIFACTORY_USER }}" >> .credentials
+          echo "CONAN_REMOTE_TOKEN=${{ secrets.BDC_ARTIFACTORY_PAT }}" >> .credentials
+      
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v3
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Build Devenv Local Version
+        uses: docker/build-push-action@v5
+        with:
+          pull: true
+          outputs: |
+            type=docker,dest=./localimage.tar
+          file: ./.devcontainer/localversion/Dockerfile
+          context: .
+          platforms: ${{ matrix.platform }}
+          tags: devcontainer/cpp:local
+
+      - name: Patch repository contents
+        run: |
+          rm -f .credentials
+          sed -i 's/ghcr.io\/eclipse-velocitas\/devcontainer-base-images\/cpp:v.*/devcontainer\/cpp:local/g' .devcontainer/Dockerfile
+          sed -i 's/sudo rm -rf ~\/.velocitas//g' .devcontainer/scripts/onCreateCommand.sh
+          sed -i 's/velocitas init//g' .devcontainer/scripts/onCreateCommand.sh
+          sed -i 's/velocitas upgrade --dry-run//g' .devcontainer/scripts/onCreateCommand.sh
+          sed -i 's/ghcr.io\/eclipse-velocitas\/devcontainer-base-images\/cpp:v.*/devcontainer\/cpp:local/g' app/Dockerfile
+
+      - name: Create archive string name
+        id: replace_string
+        run: |
+          PLATFORM=${{ matrix.platform }}
+          PLATFORM=${PLATFORM//\//_}
+          echo "printable_platform=${PLATFORM}" >> "$GITHUB_OUTPUT" 
+
+      - name: Archive repository state
+        uses: actions/upload-artifact@v4
+        with: 
+          name: devenv-local-${{ steps.replace_string.outputs.printable_platform }}
+          path: ./
