+++ app/src/CMakeLists.txt
@@ -17,8 +17,6 @@ set(TARGET_NAME "app")
 add_executable(${TARGET_NAME}
     SampleApp.cpp
     Launcher.cpp
-    FeatureManager.cpp
-    CloudNotifier.cpp
 )
 
 include_directories(${TARGET_NAME}
