+++ app/src/CloudNotifier.h
@@ -14,19 +14,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-// TODO: Add header file for CloudNotifier
-#include "CloudNotifier.h"
+#include "cloud/CloudNotifier.h"
+namespace example {
+class CloudNotifier : public virtual ::nevonex::cloud::CloudNotifier {
+public:
+    void handleMessage(const std::string& message) {
+        APP_LOG(SeverityLevel::info) << "Cloud Message Received ... " << message;
+    }
 
-#include <gtest/gtest.h>
-
-using namespace example;
-
-TEST(CloudNotifierTest, test_handleMessage_arbitraryMessage_unableToFindFeatureConfig) {
-    CloudNotifier notifier;
-    EXPECT_THROW(notifier.handleMessage(std::string{"my message"}), std::runtime_error);
-}
-
-TEST(CloudNotifierTest, handleFile_invalidFilePath_unableToFindFeatureConfig) {
-    CloudNotifier notifier;
-    EXPECT_THROW(notifier.handleFile(std::string{"/file/path"}), std::runtime_error);
-}
+    void handleFile(const ::nevonex::resource::FilePath& _file) {
+        APP_LOG(SeverityLevel::info)
+            << "Cloud File Received ...  file path : " << _file.get().string();
+    }
+};
+} // namespace example
