--- cmake/aarch64_to_x86_64.cmake
@@ -1,20 +1,0 @@
-# the name of the target operating system
-set(CMAKE_SYSTEM_NAME Linux)
-set(CMAKE_SYSTEM_PROCESSOR x86_64)
-
-set(TARGET_HOST x86_64-linux-gnu)
-
-# which compilers to use for C and C++
-set(CMAKE_C_COMPILER   ${TARGET_HOST}-gcc)
-set(CMAKE_CXX_COMPILER ${TARGET_HOST}-g++)
-
-# where is the target environment located
-set(CMAKE_FIND_ROOT_PATH  /usr/bin/x86_64-linux-gnu)
-
-# adjust the default behavior of the FIND_XXX() commands:
-# search programs in the host environment
-set(CMAKE_FIND_ROOT_PATH_MODE_PROGRAM NEVER)
-
-# search headers and libraries in the target environment
-set(CMAKE_FIND_ROOT_PATH_MODE_LIBRARY ONLY)
-set(CMAKE_FIND_ROOT_PATH_MODE_INCLUDE ONLY)
